ProjectMeetings v0.0.3 ProjectMeetings.User

This is the User module. It helps with validating users before modifying their values in Firebase.

Functions

changeset(user, params)

Validates the proposed attributes and creates a new User changeset

changeset_update(old_user, params, tokens)

Validates the proposed attributes and updates a User changeset

create_invite!(u_id, email, meeting)

Create a specified invite for the user

create_meeting!(u_id, email, meeting)

Create a specified meeting for the user

delete_invite!(u_id, email, m_id)

Remove a specified invite for the user

delete_meeting!(u_id, email, m_id)

Remove a specified meeting for the user

get_by_email(email)

Given a valid email, retrieves a user from Firebase

get_by_email!(email)

Given a valid email, retrieves a user from Firebase

get_by_firebase_token(token)

Given a valid Firebase JWT, retrieves a User from Firebase

get_by_u_id(u_id)

Given a valid u_id, retrieves a user from Firebase

get_by_u_id!(u_id)

Given a valid u_id, retrieves a user from Firebase

put(user)

Given a valid User, puts a User in Firebase