ProjectMeetings v0.0.3 ProjectMeetingsWeb.Router.Helpers

Module with named helpers generated from ProjectMeetingsWeb.Router.

Functions

meeting_path(conn_or_endpoint, action)

meeting_path(conn_or_endpoint, action, params)

meeting_path(conn_or_endpoint, action, m_id, params)

meeting_url(conn_or_endpoint, action)

meeting_url(conn_or_endpoint, action, params)

meeting_url(conn_or_endpoint, action, m_id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, email, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, email, params)