ProjectMeetings v0.0.3 ProjectMeetingsWeb.UserChannel

The user will connect to this channel to get realtime updates about invites and upcoming meetings.

Functions

code_change(old, socket, extra)

handle_in(event, message, socket)

handle_info(message, state)

Tracks the user’s presence

handle_out(binary, map, socket)

Checks if the user is online, then notifies them about an upcoming meeting accordingly

join(arg, params, socket)

Validates the user and joins the specified user channel

terminate(reason, socket)